Move endpoint classes back into *.zuul package
